voice/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
numpy==1.26.4
webrtcvad==2.0.10
websockets==12.0
openai-whisper==20240930
requests==2.32.3
soundfile==0.12.1
piper-tts==1.2.0
